move configuration routines to shared header
